+++ src/app/system/system.module.ts
@@ -9,6 +9,8 @@ import { SystemRoutingModule } from './system-routing.module';
 import { SystemComponent } from './system.component';
 import { RolesAndPermissionsComponent } from './roles-and-permissions/roles-and-permissions.component';
 import { AddRoleComponent } from './roles-and-permissions/add-role/add-role.component';
+import { EditRoleComponent } from './roles-and-permissions/edit-role/edit-role.component';
+import { ViewRoleComponent } from './roles-and-permissions/view-role/view-role.component';
 import { AuditTrailsComponent } from './audit-trails/audit-trails.component';
 import { ViewAuditComponent } from './audit-trails/view-audit/view-audit.component';
 
@@ -22,6 +24,8 @@ import { ViewAuditComponent } from './audit-trails/view-audit/view-audit.compone
         SystemComponent,
         RolesAndPermissionsComponent,
         AddRoleComponent,
+        EditRoleComponent,
+        ViewRoleComponent,
         AuditTrailsComponent,
         ViewAuditComponent
     ]