+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
         * [CP-3621] - Matomo integration
         * [CP-3641] - Updgrade to Angular 16
         * [CP-3750] - Rename Paygop Paybill to Fineract Paybill
+        * [CP-3281] - Sentry Integration
         * [CP-3662] - Add Keycloak authentication
 
 ## Version 1.1.7